refactor: 아바타 세팅 패널 재사용을 위해 분리
