+++ src/features/edit-profile-avatar/index.ts
@@ -1,6 +1,2 @@
-export { default as SelectedAvatarStateProvider } from './lib/selected-avatar-state.provider';
-
-export { default as AvatarBodyList } from './ui/avatar-body-list.component';
-export { default as AvatarFaceList } from './ui/avatar-face-list.component';
-export { default as SelectedAvatar } from './ui/selected-avatar.component';
+export { default as ProfileAvatarEditPanel } from './ui/profile-avatar-edit-panel.component';
 export { default as AvatarEditDone } from './ui/done.component';
